requirements.txt
colorama==0.4.6
coverage==7.9.1
iniconfig==2.1.0
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pluggy==1.6.0
Pygments==2.19.1
pytest==8.4.1
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
